chore: move ui to web-ui
