+++ apps/docs/next.config.js
@@ -1,4 +1,4 @@
-const withTM = require("next-transpile-modules")(["ui"]);
+const withTM = require("next-transpile-modules")(["web-ui"]);
 
 module.exports = withTM({
   reactStrictMode: true,
