+++ apps/docs/package.json
@@ -12,7 +12,7 @@
     "next": "12.3.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "ui": "workspace:*"
+    "web-ui": "workspace:*"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
