+++ packages/web-ui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ui",
+  "name": "web-ui",
   "version": "0.0.0",
   "main": "./index.tsx",
   "types": "./index.tsx",
